+++ .gitignore
@@ -9,7 +9,7 @@
 /coverage
 
 # production
-/dist
+dist
 
 # misc
 .DS_Store
