+++ src/App.tsx
@@ -1,14 +1,14 @@
 import React, { useRef } from 'react';
 import { css } from '@emotion/react';
-import CardNumberFieldset from './components/molecules/CardNumberFieldset';
+import CardNumberFieldset from './components/card-form/card-number/CardNumberFieldset';
 import AppProvider from './components/Provider';
-import CardOwnerNameFieldset from './components/molecules/CardOwnerNameFieldset';
-import ExpiredPeriodFieldset from './components/molecules/ExpiredPeriodFieldset';
-import CVCFieldset from './components/molecules/CVCFieldset';
-import CardPasswordFieldset from './components/molecules/CardPasswordFieldset';
-import Navigation from './components/molecules/Navigation';
-import ConfirmButtonContainer from './components/atoms/confirm-button/ConfirmButtonContainer';
-import CardContainer from './components/atoms/card/CardContainer';
+import CardOwnerNameFieldset from './components/card-form/card-owner/CardOwnerNameFieldset';
+import ExpiredPeriodFieldset from './components/card-form/card-expired-period/ExpiredPeriodFieldset';
+import CVCFieldset from './components/card-form/card-cvc/CVCFieldset';
+import CardPasswordFieldset from './components/card-form/card-password/CardPasswordFieldset';
+import Navigation from './components/navigation/Navigation';
+import ConfirmButtonContainer from './components/card-form/confirm-button/ConfirmButtonContainer';
+import CardContainer from './components/card/CardContainer';
 
 function App() {
   return (
