+++ src/components/card-form/card-expired-period/ExpiredPeriodFieldset.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import ExpiredPeriodInputContainer from '../atoms/card-expired-period/ExpiredPeriodInputContainer';
-import { Body, Fieldset, Head } from '../templates/Fieldset';
+import ExpiredPeriodInputContainer from '../card-expired-period/ExpiredPeriodInputContainer';
+import { Body, Fieldset, Head } from '../Fieldset';
 
 function ExpiredPeriodFieldset() {
   return (
